Add(seokho/boards): 게시글 수정 삭제 요청자 검증 추가 #288

+++ app/src/models/services/Board/Board.js
@@ -4,6 +4,7 @@ const BoardStorage = require('./BoardStorage');
 const Notification = require('../Notification/Notification');
 const NotificationStorage = require('../Notification/NotificationStorage');
 const Error = require('../../utils/Error');
+const WiterCheck = require('../../utils/WriterCheck');
 const boardCategory = require('../Category/board');
 
 class Board {
@@ -203,6 +204,14 @@ class Board {
         boardNum: params.boardNum,
       };
 
+      const writerCheck = await WiterCheck.ctrl(
+        this.auth.id,
+        boardInfo.boardNum,
+        'boards'
+      );
+
+      if (!writerCheck.success) return writerCheck;
+
       const updateBoardCnt = await BoardStorage.updateOneByBoardNum(boardInfo);
 
       if (updateBoardCnt === 0) {
@@ -223,6 +232,14 @@ class Board {
         category,
         boardNum: params.boardNum,
       };
+
+      const writerCheck = await WiterCheck.ctrl(
+        this.auth.id,
+        boardInfo.boardNum,
+        'boards'
+      );
+
+      if (!writerCheck.success) return writerCheck;
 
       const deleteBoardCnt = await BoardStorage.deleteOneByBoardNum(boardInfo);
